+++ R/bookmarks.R
@@ -85,10 +85,7 @@ get_bookmarks <- function(filename, use_names = TRUE) {
     if (supports_pdftk()) {
         get_bookmarks_pdftk(filename, use_names = use_names)
     } else {
-        msg <- c(need_to_install_str("get_bookmarks()"),
-                 install_pdftk_str()
-        )
-        abort(msg, class = "xmpdf_suggested_package")
+        abort(msg_get_bookmarks(), class = "xmpdf_suggested_package")
     }
 }
 
@@ -286,11 +283,7 @@ set_bookmarks <- function(bookmarks, input, output = input) {
     } else if (supports_pdftk()) {
         set_bookmarks_pdftk(bookmarks, input, output)
     } else {
-        msg <- c(need_to_install_str("set_bookmarks()"),
-                 install_gs_str(),
-                 install_pdftk_str()
-        )
-        abort(msg, class = "xmpdf_suggested_package")
+        abort(msg_set_bookmarks(), class = "xmpdf_suggested_package")
     }
 }