+++ R/edit_xmp.R
@@ -53,10 +53,7 @@ get_xmp <- function(filename, use_names = TRUE) {
     if (supports_exiftool()) {
         get_xmp_exiftool(filename, use_names = use_names)
     } else {
-        msg <- c(need_to_install_str("get_xmp()"),
-                 install_exiftool_str()
-        )
-        abort(msg, class = "xmpdf_suggested_package")
+        abort(msg_get_xmp(), class = "xmpdf_suggested_package")
     }
 }
 
@@ -104,10 +101,7 @@ set_xmp <- function(xmp, input, output = input) {
     if (supports_exiftool()) {
         set_xmp_exiftool(xmp, input, output)
     } else {
-        msg <- c(need_to_install_str("set_xmp()"),
-                 install_exiftool_str()
-        )
-        abort(msg, class = "xmpdf_suggested_package")
+        abort(msg_set_xmp(), class = "xmpdf_suggested_package")
     }
 }